requirements.txt
Django==3.1.1
djangorestframework==3.11.1
django-extensions==3.0.9
selenium==3.141.0
requests==2.24.0
nodeenv==1.5.0
